feat: single content type support
