--- services/constant.js
@@ -1,3 +1,0 @@
-module.exports = {
-    TEMPLATE_DEFAULT: 'Generic',
-};
