+++ services/utils/constant.js
@@ -1,0 +1,10 @@
+module.exports = {
+    TEMPLATE_DEFAULT: 'Generic',
+    KIND_TYPES: {
+        SINGLE: 'singleType',
+        COLLECTION: 'collectionType'
+    },
+    MODEL_TYPES: {
+        CONTENT_TYPE: 'contentType'
+    },
+};
